fix getCount
