+++ Modules/Utils.cs
@@ -2759,7 +2759,6 @@ public static class Utils
     }
     public static void AfterPlayerDeathTasks(PlayerControl target, bool onMeeting = false)
     {
-        // CustomRoleSelector.GhostAssignPatch(target); should be handled in exiled() patch, if not we remove commenting.
 
         switch (target.GetCustomRole())
         {
